// src/types/enums.ts
export enum ModalTypesEnum {
  SIGN_UP = "SIGN_UP",
  LOG_IN = "LOG_IN",
  DISCUSSION = "DISCUSSION",
  CREATE_POST = "CREATE_POST",
  EDIT_POST = "EDIT_POST",
}

export enum BtnVariantEnum {
  PRIMARY = "PRIMARY",
  SECONDARY = "SECONDARY",
  SHADOW_PRIMARY = "SHADOW_PRIMARY",
  SHADOW_SECONDARY = "SHADOW_SECONDARY",
  SUCCESS = "SUCCESS",
  SECONDARY_BLUE = "SECONDARY_BLUE",
}

export enum ContainerVariantEnum {
  FULL = "FULL",
  PARTIAL = "PARTIAL",
}

export enum TextAreaVariantEnum {
  NORMAL = "NORMAL",
  SHADOW = "SHADOW",
}

export enum CardTypeVariantEnum {
  NORMAL = "NORMAL",
  POST = "POST",
}

export enum SelectTypeVariantEnum {
  NORMAL = "NORMAL",
  CHECKED = "CHECKED",
}

export enum PostComponentTypeEnum {
  MARK_CLAIM = "MARK_CLAIM",
  VOTING_STATS = "VOTING_STATS",
}

export enum SwalTypeEnum {
  SUCCESS = "success",
  ERROR = "error",
  WARNING = "warning",
}